requirements.txt
asgiref==3.7.2
Django==4.2.4
djangorestframework==3.14.0
psycopg2==2.9.7
python-ffmpeg-video-streaming==0.1.15
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
